terraviva/backend/requirements/dev.txt
-r base.txt

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-xdist==3.5.0
factory-boy==3.3.1

# Code Quality
ruff==0.9.3

# Type Checking
mypy==1.14.1
django-stubs==5.1.1
django-types
